Refactoring: Move id helper to model module
